what I want to work‣

+++ app/models/year.rb
@@ -1,6 +1,6 @@
 class Year 
   include Neo4j::ActiveNode
-  property :number, type: Integer
+  property :number, type: Float
 
   has_many :out, :categories, type: :CAT, unique: true
 end
